refactor: main -> lib
