+++ src/lib.rs
@@ -1,0 +1,11 @@
+pub mod cli;
+pub mod engine;
+pub mod facts;
+pub mod progress;
+pub mod project;
+pub mod repo;
+pub mod snapshot;
+pub mod stash;
+pub mod theme;
+
+include!(concat!(env!("OUT_DIR"), "/built.rs"));
